Fix surface display on upload and conformance crash on large meshes
Root causes:
1. Surfaces only appeared after conformance because upload handler converted
   flat typed arrays to Vec3 object arrays (millions of {x,y,z} objects),
   which was slow and unnecessary.
2. Conformance crashed at "Transferring results" because serde_wasm_bindgen
   serialized millions of Vec3 structs as individual JS objects, causing OOM.

Fixes:
- Add run_conformance_flat() and parse_surface_flat() WASM functions that
  return Float32Array/Uint32Array directly from Rust, bypassing serde for
  mesh data (summary/metadata still uses serde).
- Store UploadedSurface with flat typed arrays (Float32Array positions,
  Uint32Array indices) instead of Vec3[]/[number,number,number][].
- Worker stores raw binary data and uses new flat WASM functions.
- Viewer reads positions/indices directly from UploadedSurface — no
  intermediate object array conversion.
- Surfaces render immediately on upload, before any conformance.
- ConformanceResult for charts/reports carries only volume/summary metadata,
  no mesh vertex data (display uses FlatDomainSolid typed arrays).
- Cross-section converts to Vec3[] lazily on demand only.
- Add 2M-triangle large surface test (cargo test --release, 6s).

+++ web/src/workers/engineClient.ts
@@ -4,7 +4,7 @@ export interface FlatSurface {
   name: string;
   role: SurfaceRole;
   fileName: string;
-  positions: Float64Array;
+  positions: Float32Array;
   indices: Uint32Array;
   vertexCount: number;
   triangleCount: number;
@@ -16,7 +16,7 @@ export interface FlatDomainSolid {
   color: string;
   volume: number;
   block_name?: string;
-  positions: Float64Array;
+  positions: Float32Array;
   indices: Uint32Array;
   vertexCount: number;
   triangleCount: number;
